feat: 썸원 이벤트 대응

+++ photo-service/src/main/java/kr/mafoo/photo/controller/dto/response/SumoneSummaryResponse.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "리캡 응답")
 public record SumoneSummaryResponse(
         @Schema(description = "사용자 수", example = "1565")
-        int userCount
+        Long userCount
 ){
 
 }
